Solution for 127. Word Ladder.
-Solution implemented using two way search.
-In worest case, this implementaion works as a one ways search.
-In best cenaria this will same a lot of time because optimaztion will
choose always smeller set and work on that part. We will go with the
greedy way and avoid the way where is the big increase of the possible
ways.
-This is where clever idea(of course not mine I am refering discussion
sol) because it's also deleting visited words from dic and adding to
another set and at that way we will removing word over head.